customer update

+++ app/Http/Controllers/MeasureController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Customer;
 use App\Models\Measure;
 use Illuminate\Http\Request;
-
+use App\Http\Resources\MeasureResource;
 use Illuminate\Support\Facades\Auth;
 
 class MeasureController extends Controller
@@ -13,57 +13,51 @@ class MeasureController extends Controller
 
     public function __construct()
     {
-        // $this->middleware('auth:sanctum');
-        // $this->middleware('auth:sanctum');
+        $this->middleware('auth:sanctum');
     }
 
     public function store(Request $request)
     {
-        //$customer = Auth::user();
+        $customer = Auth::user();
         $name = $request->validate([
-            //'customer_id' =>'required|integer|exists:customers,id',
             'name' => 'required|string|max:255'
-            
         ]);
-
-        //$customer_id = $customer->id;
 
         $newMeasure = Measure::create([
             'name' => $name['name'],
-            // 'customer_id' => $customer->id
-            //'customer_id' => '11'
+            'customer_id' => $customer->id
         ]);
 
         return response()->json([
             'message' => 'measure created successfully',
-            "data" => $newMeasure
+            "data" => new MeasureResource($newMeasure)
         ]);
     }
 
 
-     public function show(Measure $measure)
+    public function show(Measure $measure)
     {
-        // $customer = Auth::user();
+        $customer = Auth::user();
 
-        // if ($measure->customer_id !== $customer->id) {
-        //     return response()->json(['message' => 'Unauthorized.'], 403);
-        // }
+        if ($measure->customer_id !== $customer->id) {
+            return response()->json(['message' => 'Unauthorized.'], 403);
+        }
 
-        $measure->load('secondary_measures.measure_name'); 
+        $measure->load('secondary_measures.measure_name');
 
         return response()->json([
             'message' => 'Measure profile retrieved successfully',
-            'data' => $measure,
+            'data' =>new MeasureResource( $measure),
         ]);
     }
-    
+
     public function update(Request $request, Measure $measure)
     {
-        // $customer = Auth::user();
+        $customer = Auth::user();
 
-        // if ($measure->customer_id !== $customer->id) {
-        //     return response()->json(['message' => 'Unauthorized.'], 403);
-        // }
+        if ($measure->customer_id !== $customer->id) {
+            return response()->json(['message' => 'Unauthorized.'], 403);
+        }
 
         $validated = $request->validate([
             'name' => 'required|string|max:255',
@@ -73,24 +67,23 @@ class MeasureController extends Controller
 
         return response()->json([
             'message' => 'Measure profile updated successfully',
-            'data' => $measure,
+            'data' => new MeasureResource( $measure),
         ]);
     }
 
     public function destroy(Measure $measure)
     {
-        // $customer = Auth::user();
+        $customer = Auth::user();
 
-        // if ($measure->customer_id !== $customer->id) {
-        //     return response()->json(['message' => 'Unauthorized.'], 403);
-        // }
+        if ($measure->customer_id !== $customer->id) {
+            return response()->json(['message' => 'Unauthorized.'], 403);
+        }
 
-        $measure->secondary_measures()->delete(); // delete all related values
-        $measure->delete(); // delete the profile
+        $measure->secondary_measures()->delete(); 
+        $measure->delete();
 
         return response()->json([
             'message' => 'Measure profile and values deleted successfully',
         ]);
     }
-} 
-
+}
